+++ Wordschatz.Domain/Models/Theme.cs
@@ -8,9 +8,9 @@ namespace Wordschatz.Domain.Models
     public class Theme : EventSourcedAggregate
     {
         public string Name { get; set; }
-        public int DictionaryId { get; set; }
+        public ulong DictionaryId { get; set; }
         public virtual Dictionary Dictionary { get; set; }
-        public int? ParentId { get; set; }
+        public ulong? ParentId { get; set; }
         public virtual Theme Parent { get; set; }
 
         public virtual List<WordTheme> Words { get; set; }
